requirements.txt
keystoneauth1==4.5.0
gnocchiclient==7.0.4
PyYAML==5.4.1
Flask==2.0.1
shade==1.33.0
psutil==5.8.0
gnocchiclient==7.0.7
oslo.i18n==5.0.1
oslo.serialization==4.1.0
oslo.utils==4.8.0